bug: не работает создание комментариев
