--- js/comments.js
@@ -1,6 +1,0 @@
-const socialCommentsBlock = document.querySelector('.social__comments');
-const socialCommentsElement = socialCommentsBlock.createElement('li');
-socialCommentsBlock.appendChild(socialCommentsElement);
-const socialCommentsImg = socialCommentsElement.createElement('img');
-socialCommentsImg.classList.add('social__picture');
-// socialCommentsImg.src =
