+++ js/main.js
@@ -1,3 +1,4 @@
 import {addMiniatureFN} from './miniature.js';
 import {postsArray} from './miniature.js';
+
 addMiniatureFN(postsArray);
