+++ js/miniature.js
@@ -1,6 +1,7 @@
 import {createPostsArray} from './data.js';
 import {POSTSQUANTITY} from './data.js';
 import {showBigPicture} from './fullpicture.js';
+import {createCommentFN} from './create-comments.js';
 
 const pictureList = document.querySelector('.pictures');
 const newPictureTemplate = document.querySelector('#picture').content.querySelector('.picture');
@@ -16,11 +17,12 @@ const addMiniatureFN = (array) => {
     picture.querySelector('.picture__img').alt = description;
     picture.querySelector('.picture__likes').textContent = likes;
     picture.querySelector('.picture__comments').textContent = comments.length;
-
+    const pictureComments = comments;
     PicturesFragment.append(picture);
 
     picture.addEventListener('click', () => {
       showBigPicture(picture);
+      createCommentFN(pictureComments);
     });
   });
   pictureList.append(PicturesFragment);
